pyproject.toml
[project]
name = "marketing-video-agent-factory"
version = "1.0.0"
description = "Multi-agent marketing video generation platform powered by Google ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.13.0",
    "google-genai>=1.17.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.1.0",
    "pillow>=10.0.0",
    "colorthief>=0.2.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
